requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
Flask-Limiter==3.5.0
email-validator==2.1.0
Pillow==12.0.0 # lidando com imagens
sendgrid==6.11.0 # enviar emails